-- adding in base absence pay instruction which the other absence pay instructions can inherit from
