+++ services/payInstructions/SspPayInstruction.js
@@ -1,21 +1,5 @@
-const BaseInstruction = require("./BaseInstruction");
+const AbsencePayInstruction = require("./AbsencePayInstruction");
 const FormUtils = require("../form-utils");
 
-module.exports = class SspPayInstruction extends BaseInstruction {
-    get canInstructionsOverlap() {
-        return true;
-    }
-
-    parseForApi(body) {
-        let cleanBody = super.parseForApi(body);
-
-        if (cleanBody.StatutoryOffset && cleanBody.StatutoryOffset.toLowerCase() === "on") {
-            cleanBody.StatutoryOffset = true;
-        }
-        else {
-            cleanBody.StatutoryOffset = false;
-        }
-
-        return cleanBody;
-    }
+module.exports = class SspPayInstruction extends AbsencePayInstruction {
 };
